GDGT-2589 addressing PR's comments
